Create each API test class
